debut couleur

+++ src/main/java/graph/Color.java
@@ -1,30 +1,30 @@
 package graph;
 
 public class Color {
-	private int r,g,b;
+	private double r,g,b;
 	
-	public Color(int[] rgbTab) {
+	public Color(double[] rgbTab) {
 		this.r = rgbTab[0];
 		this.g = rgbTab[1];
 		this.b = rgbTab[2];
 	}
 
-	public Color(int i, int j, int k) {
+	public Color(double i, double j, double k) {
 		this.r = i;
 		this.g = j;
 		this.b = k;
 	}
 
 	public int getR() {
-		return r;
+		return (int)r;
 	}
 
 	public int getG() {
-		return g;
+		return (int)g;
 	}
 
 	public int getB() {
-		return b;
+		return (int)b;
 	}
 
 	@Override
